refactor(cli,ironfish): Move a few more methods
